+++ CHANGELOG.md
@@ -12,7 +12,10 @@
 
 ## Bug fixes:
 
-*Contributors of this release (in alphabetical order):* @chekkan, @Code-Grump, @konarx
+* Fix: Exceptions from before scenario plugin hook events (RuntimePluginTestExecutionLifecycleEvents.BeforeScenario) are not shown (#856)
+* Fix: Before scenario hooks are not raised immediately when `runtime/stopAtFirstError` is true, but handled delayed (#857)
+
+*Contributors of this release (in alphabetical order):* @chekkan, @Code-Grump, @gasparnagy, @konarx
 
 # v3.0.3 - 2025-09-17